Time_Series_Forecasting_Alerts_Mails/requirements.txt
pymongo
flask
flask-cors
python-dotenv
schedule
